refactor in pr package
