+++ testutil/pr/verifypr_internal_test.go
@@ -13,7 +13,7 @@
 // You should have received a copy of the GNU General Public License along with
 // this program.  If not, see <http://www.gnu.org/licenses/>.
 
-package main
+package pr
 
 import (
 	"fmt"
